+++ lib/http_cookie.mli
@@ -18,18 +18,42 @@
     The library supports consuming and creating HTTP cookie in both requests and
     responses.
 
-    The standard implemented by the library is
-    {{:https://tools.ietf.org/html/rfc6265} RFC 6265}. *)
+    The standards implemented by the library is
+
+    + {{:https://tools.ietf.org/html/rfc6265} RFC 6265 - Cookies}.
+    + {{:https://datatracker.ietf.org/doc/html/rfc1123} Section 3.3.1 - HTTP
+      Date}
+    + {{:https://datatracker.ietf.org/doc/html/rfc1034#section-3.5} Domain Name}
+    + {{:https://datatracker.ietf.org/doc/html/rfc1123#section-2.1} Hosts}
+    + {{:https://datatracker.ietf.org/doc/html/draft-main-ipaddr-text-rep-02#section-3}
+      IPv4/IPv6 Address} *)
 
 (** A HTTP cookie. *)
 type t
 
-(** Cookie expiry datetime value. *)
+(** normalized date time value in GMT. *)
 and date_time
 
-(** 'Same-site' cookie attribute. See
-    https://tools.ietf.org/html/draft-ietf-httpbis-cookie-same-site-00. *)
+(** 'Same-site' cookie attribute.
+    {{:https://tools.ietf.org/html/draft-ietf-httpbis-cookie-same-site-00}
+    Same-site} *)
 and same_site = [`None | `Lax | `Strict]
+
+(** {1 Pretty Printers} *)
+
+val pp : Format.formatter -> t -> unit
+val pp_same_site : Format.formatter -> same_site -> unit
+
+val pp_date_time : Format.formatter -> date_time -> unit
+(** [pp_date_time fmt date_time] pretty prints {!type:date_time} in RFC 1123
+    format.
+
+    Example:
+
+    {[ Sun, 06 Nov 1994 08:49:37 GMT ]} *)
+
+val pp_rfc1123 : Format.formatter -> date_time -> unit
+(** Alias of {!val:pp_date_time}. *)
 
 (** {1 Create/Decode/Encode} *)
 
@@ -49,19 +73,20 @@ val date_time :
        | `Nov
        | `Dec ]
   -> weekday:[`Sun | `Mon | `Tue | `Wed | `Thu | `Fri | `Sat]
-  -> day_of_month:int
+  -> day:int
   -> hour:int (** 24 hour format *)
   -> minutes:int
   -> seconds:int
   -> (date_time, string) result
 (** [date_time] is [Ok dt] if all of the given parameters are valid for creating
-    {!type:date_time} value, otherwise [Error err] is denotes the error. *)
+    {!type:date_time} value, otherwise it is [Error err] where err denotes the
+    error. *)
 
 val create :
      ?path:string
   -> ?domain:string
   -> ?expires:date_time
-  -> ?max_age:int
+  -> ?max_age:int64
   -> ?secure:bool
   -> ?http_only:bool
   -> ?same_site:same_site
@@ -109,6 +134,15 @@ val to_set_cookie : t -> string
 SID=31d4d96e407aad42; Path=/; Secure; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT
     v} *)
 
+val of_set_cookie : string -> (t, string) result
+(** [of_set_cookie s] is [Ok cookie] if [s] can be parsed successfully to create
+    {!type:t}.
+
+    [s] is the HTTP 'Set-Cookie' header value. The syntax for the value is
+    defined as [set-cookie-string] in
+    {{:https://datatracker.ietf.org/doc/html/rfc6265#section-4.1.1} RFC 6265,
+    4.1} *)
+
 (** {1 Cookie Attributes}
 
     Cookie attributes are defined precisely at
@@ -139,18 +173,18 @@ val expires : t -> date_time option
 
     See {{:https://tools.ietf.org/html/rfc6265#section-4.1.2.1.} cookie-expires} *)
 
-val max_age : t -> int option
+val max_age : t -> int64 option
 (** [max_age t] returns a cookie max_age attribute.
 
     See {{:https://tools.ietf.org/html/rfc6265#section-4.1.2.2} max-age} and
     {{:https://tools.ietf.org/html/rfc6265#section-4.1.1} max-age-av} *)
 
-val secure : t -> bool option
+val secure : t -> bool
 (** [secure t] returns a secure attribute.
 
     See {{:https://tools.ietf.org/html/rfc6265#section-4.1.2.5} cookie-secure} *)
 
-val http_only : t -> bool option
+val http_only : t -> bool
 (** [http_only t] returns a http_only attribute.
 
     See {{:https://tools.ietf.org/html/rfc6265#section-4.1.2.6} http-only} *)
@@ -182,14 +216,8 @@ val update_name : string -> t -> (t, string) result
 val update_path : string option -> t -> (t, string) result
 val update_domain : string option -> t -> (t, string) result
 val update_expires : date_time option -> t -> t
-val update_max_age : int option -> t -> (t, string) result
-val update_secure : bool option -> t -> t
-val update_http_only : bool option -> t -> t
+val update_max_age : int64 option -> t -> (t, string) result
+val update_secure : bool -> t -> t
+val update_http_only : bool -> t -> t
 val update_same_site : same_site option -> t -> t
 val update_extension : string option -> t -> (t, string) result
-
-(** {1 Pretty Printers} *)
-
-val pp : Format.formatter -> t -> unit
-val pp_date_time : Format.formatter -> date_time -> unit
-val pp_same_site : Format.formatter -> same_site -> unit